Aichallenge/requirements.txt
langchain
chromadb
gradio
crewai
sentence-transformers
PyPDF2
python-docx
transformers
torch

reportlab
Pillow
pytesseract
